+++ python/GafferRenderManUI/RenderManMeshLightUI.py
@@ -34,14 +34,8 @@
 #
 ##########################################################################
 
-import functools
-
 import Gaffer
 import GafferRenderMan
-
-def __shaderMetadata( plug, name ) :
-
-	return Gaffer.Metadata.value( plug.node()["__shader"].descendant( plug.relativeName( plug.node() ) ), name )
 
 Gaffer.Metadata.registerNode(
 
@@ -65,6 +59,7 @@ Gaffer.Metadata.registerNode(
 			""",
 
 			"nameValuePlugPlugValueWidget:ignoreNamePlug" : True,
+			"layout:index" : 0,
 
 		},
 
@@ -76,48 +71,9 @@ Gaffer.Metadata.registerNode(
 
 		"parameters" : {
 
-			"description" :
-			"""
-			The parameters of the PxrMeshLight shader that is applied to the
-			meshes.
-			""",
-
-			## \todo Extend the Metadata API so we can register a provider for "*",
-			# which can automatically transfer all internal metadata.
-			"noduleLayout:section" : functools.partial( __shaderMetadata, name = "noduleLayout:section" ),
-			"nodule:type" : functools.partial( __shaderMetadata, name = "nodule:type" ),
-			"noduleLayout:spacing" : functools.partial( __shaderMetadata, name = "noduleLayout:spacing" ),
-			"plugValueWidget:type" : functools.partial( __shaderMetadata, name = "plugValueWidget:type" ),
-			"layout:section:Basic:collapsed" : False
+			"layout:section:Basic:collapsed" : False,
 
 		},
-
-		"parameters.*" : {
-
-			"description" : functools.partial( __shaderMetadata, name = "description" ),
-			"nodule:type" : functools.partial( __shaderMetadata, name = "nodule:type" ),
-			"noduleLayout:section" : functools.partial( __shaderMetadata, name = "noduleLayout:section" ),
-			"nodule:color" : functools.partial( __shaderMetadata, name = "nodule:color" ),
-			"layout:section" : functools.partial( __shaderMetadata, name = "layout:section" ),
-			"plugValueWidget:type" : functools.partial( __shaderMetadata, name = "plugValueWidget:type" ),
-			"presetNames" : functools.partial( __shaderMetadata, name = "presetNames" ),
-			"presetValues" : functools.partial( __shaderMetadata, name = "presetValues" ),
-
-		},
-
-		"defaultLight" : {
-
-			"description" :
-			"""
-			Whether this light illuminates all geometry by default. When
-			toggled, the light will be added to the `defaultLights` set, which
-			can be referenced in set expressions and manipulated by downstream
-			nodes.
-			""",
-
-			"layout:section" : "Light Linking",
-
-		}
 
 	}